update namespace for newsletter
