+++ app/graphql/mutations/newsletter/subscribe_user_to_newsletter_mutation.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Mutations
+  module Newsletter
+    class SubscribeUserToNewsletterMutation < BaseMutation
+      argument :input, Types::Custom::Inputs::Mutations::SubscribeUserToNewsletterInput, required: true
+      type Types::Custom::Objects::UserObject
+  
+      def resolve(params)
+        super(params)
+        ::Newsletters::SubscribeToNewsletterService.call(params: @params)
+      end
+    end
+  end
+end
