--- app/graphql/mutations/unsubscribe_user_from_newsletter_mutation.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module Mutations
-  class UnsubscribeUserFromNewsletterMutation < GraphQL::Schema::Mutation
-    argument :email, String, required: true
-    type Types::Custom::Objects::UserObject
-
-    def resolve(params)
-      ::Newsletters::UnsubscribeFromNewsletterService.call(email: params.fetch(:email))
-    end
-  end
-end
